Split the old call into seperated runtime file
