+++ node/runtime/pangolin/src/bridges/substrate/pangoro_messages.rs
@@ -29,6 +29,9 @@ use bridge_primitives::{
 	DarwiniaFromThisChainMessageVerifier, PANGOLIN_CHAIN_ID, PANGORO_CHAIN_ID,
 	PANGORO_PANGOLIN_LANE, WITH_PANGOLIN_MESSAGES_PALLET_NAME,
 };
+use darwinia_support::to_bytes32;
+use dp_asset::token::Token;
+use dp_s2s::{CallParams, EncodeCall};
 
 /// Message payload for Pangolin -> Pangoro messages.
 pub type ToPangoroMessagePayload = FromThisChainMessagePayload<WithPangoroMessageBridge>;
@@ -252,5 +255,44 @@ impl SourceHeaderChain<pangoro_primitives::Balance> for Pangoro {
 			proof,
 			messages_count,
 		)
+	}
+}
+
+/// The bridged chain(Pangoro) dispatch info
+#[derive(Encode, Decode, Debug, PartialEq, Eq, Clone)]
+pub enum PangoroRuntime {
+	/// NOTE: The index must be the same as the backing pallet in the pangoro runtime
+	#[codec(index = 20)]
+	Sub2SubBacking(PangoroSub2SubBackingCall),
+}
+
+/// The backing call in the pangoro s2s backing pallet
+#[derive(Encode, Decode, Debug, PartialEq, Eq, Clone)]
+#[allow(non_camel_case_types)]
+pub enum PangoroSub2SubBackingCall {
+	/// NOTE: The index depends on the call order in the s2s backing pallet.
+	#[codec(index = 2)]
+	unlock_from_remote(Token, AccountId),
+}
+
+pub struct PangoroRuntimeCalls;
+impl EncodeCall<AccountId> for PangoroRuntimeCalls {
+	fn encode_call(call_params: CallParams<AccountId>) -> Result<Vec<u8>, ()> {
+		let call = match call_params {
+			CallParams::UnlockFromRemote(_account_id, unlock_info) => {
+				if unlock_info.recipient.len() != 32 {
+					return Err(());
+				}
+
+				let recipient_id: AccountId = to_bytes32(unlock_info.recipient.as_slice()).into();
+				PangoroRuntime::Sub2SubBacking(PangoroSub2SubBackingCall::unlock_from_remote(
+					unlock_info.token,
+					recipient_id,
+				))
+				.encode()
+			}
+			_ => return Err(()),
+		};
+		Ok(call)
 	}
 }
